% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_interactive.R
\name{annotate_interactive}
\alias{annotate_interactive}
\title{interactive annotations}
\usage{
annotate_interactive(...)
}
\arguments{
\item{...}{arguments passed to base geometry.}
}
\description{
The geometry is based on \code{\link[ggplot2]{annotate}}.
See the documentation for those functions for more details.
}
\examples{
# add interactive annotation -------
library(ggplot2)
library(ggiraph)

gg <- ggplot(mtcars, aes(x = disp, y = qsec )) +
  geom_point(size=2) +
  annotate_interactive(
    "rect", xmin = 100, xmax = 400, fill = "red",
    data_id = "an_id", tooltip = "a tooltip",
    ymin = 18, ymax = 20, alpha = .5)

x <- girafe(ggobj = gg, width_svg = 5, height_svg = 4)
if( interactive() ) print(x)
}
\seealso{
\code{\link{girafe}}
}
