% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_polygon_grob.R
\name{drawDetails.interactive_polygon_grob}
\alias{drawDetails.interactive_polygon_grob}
\title{interactive_polygon_grob drawing}
\usage{
\method{drawDetails}{interactive_polygon_grob}(x, recording)
}
\arguments{
\item{x}{ A grid grob. }

\item{recording}{ A logical value indicating whether a grob
  is being added to the display list or redrawn from the display list. }
}
\description{
draw an interactive_polygon_grob
}

