% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_segments_grob.R
\name{drawDetails.interactive_segments_grob}
\alias{drawDetails.interactive_segments_grob}
\title{interactive_segments_grob drawing}
\usage{
\method{drawDetails}{interactive_segments_grob}(x, recording)
}
\arguments{
\item{x}{ A grid grob. }

\item{recording}{ A logical value indicating whether a grob
  is being added to the display list or redrawn from the display list. }
}
\description{
draw an interactive_segments_grob
}
