% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactivePointsGrob.R
\name{drawDetails.interactivePointsGrob}
\alias{drawDetails.interactivePointsGrob}
\title{interactivePointsGrob drawing}
\usage{
\method{drawDetails}{interactivePointsGrob}(x, recording)
}
\arguments{
\item{x}{ A grid grob. }

\item{recording}{ A logical value indicating whether a grob
  is being added to the display list or redrawn from the display list. }
}
\description{
draw an interactivePointsGrob
}

