% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactivePolygonGrob.R
\name{drawDetails.interactivePolygonGrob}
\alias{drawDetails.interactivePolygonGrob}
\title{interactivePolygonGrob drawing}
\usage{
\method{drawDetails}{interactivePolygonGrob}(x, recording)
}
\arguments{
\item{x}{ A grid grob. }

\item{recording}{ A logical value indicating whether a grob
  is being added to the display list or redrawn from the display list. }
}
\description{
draw an interactivePolygonGrob
}

