% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactivePolylineGrob.R
\name{drawDetails.interactivePolylineGrob}
\alias{drawDetails.interactivePolylineGrob}
\title{interactivePolylineGrob drawing}
\usage{
\method{drawDetails}{interactivePolylineGrob}(x, recording)
}
\arguments{
\item{x}{ A grid grob. }

\item{recording}{ A logical value indicating whether a grob
  is being added to the display list or redrawn from the display list. }
}
\description{
draw an interactivePolylineGrob
}

