% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveRectGrob.R
\name{drawDetails.interactiveRectGrob}
\alias{drawDetails.interactiveRectGrob}
\title{interactiveRectGrob drawing}
\usage{
\method{drawDetails}{interactiveRectGrob}(x, recording)
}
\arguments{
\item{x}{ A grid grob. }

\item{recording}{ A logical value indicating whether a grob
  is being added to the display list or redrawn from the display list. }
}
\description{
draw an interactiveRectGrob
}

