% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveSegmentsGrob.R
\name{drawDetails.interactiveSegmentsGrob}
\alias{drawDetails.interactiveSegmentsGrob}
\title{interactiveSegmentsGrob drawing}
\usage{
\method{drawDetails}{interactiveSegmentsGrob}(x, recording)
}
\arguments{
\item{x}{ A grid grob. }

\item{recording}{ A logical value indicating whether a grob
  is being added to the display list or redrawn from the display list. }
}
\description{
draw an interactiveSegmentsGrob
}

