% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracers.R
\name{dsvg_tracer_off}
\alias{dsvg_tracer_off}
\title{trace off id colection}
\usage{
dsvg_tracer_off()
}
\value{
graphical elements id as integer values.
}
\description{
get collected id of an dsvg device and
stop collecting.
}
