% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsvg_view.R
\name{dsvg_view}
\alias{dsvg_view}
\title{Run plotting code and view svg in RStudio Viewer or web broswer.}
\usage{
dsvg_view(code, ...)
}
\arguments{
\item{code}{Plotting code to execute.}

\item{...}{Other arguments passed on to \code{\link{dsvg}}.}
}
\description{
This is useful primarily for testing. Requires the \code{htmltools}
package.
}
\examples{
\donttest{
dsvg_view(plot(1:10))
dsvg_view(hist(rnorm(100)))
}
}
