% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_interactive.R
\name{facet_wrap_interactive}
\alias{facet_wrap_interactive}
\title{Create interactive wraped facets}
\usage{
facet_wrap_interactive(...)
}
\arguments{
\item{...}{arguments passed to base function and
\code{\link[=labeller_interactive]{labeller_interactive()}} for argument \code{labeller}.}
}
\value{
An interactive facetting object.
}
\description{
These facets are based on
\code{\link[=facet_wrap]{facet_wrap()}}.

To make a facet interactive, it is mandatory to use
\code{\link[=labeller_interactive]{labeller_interactive()}} for argument \code{labeller}.
}
\seealso{
\code{\link[=girafe]{girafe()}}
}
