% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{font_family_exists}
\alias{font_family_exists}
\title{Check if font family exists.}
\usage{
font_family_exists(font_family = "sans")
}
\arguments{
\item{font_family}{font family name (case sensitive)}
}
\value{
A logical value
}
\description{
Check if a font family exists in system fonts.
}
\examples{
font_family_exists("sans")
font_family_exists("Arial")
font_family_exists("Courier")
}
