% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_boxplot_interactive.R
\name{geom_boxplot_interactive}
\alias{geom_boxplot_interactive}
\title{interactive boxplot}
\usage{
geom_boxplot_interactive(...)
}
\arguments{
\item{...}{arguments passed to base geometry.}
}
\description{
The geometry is based on \code{\link[ggplot2]{geom_boxplot}}.
See the documentation for those functions for more details.
}
\examples{
# add interactive boxplot -------
library(ggplot2)
library(ggiraph)

p <- ggplot(mpg,
  aes(x = class, y = hwy, tooltip = class)) +
  geom_boxplot_interactive()

x <- girafe(ggobj = p)
if( interactive() ) print(x)

p <- ggplot(mpg, aes(x = drv, y = hwy, tooltip = class, fill = class, data_id=class)) +
  geom_boxplot_interactive(outlier.colour = "red") +
  guides(fill = "none") + theme_minimal()

x <- girafe(ggobj = p)
if( interactive() ) print(x)
}
\seealso{
\code{\link{girafe}}
}
