% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_dotplot_interactive.R
\name{geom_dotplot_interactive}
\alias{geom_dotplot_interactive}
\title{Create interactive dot plots}
\usage{
geom_dotplot_interactive(...)
}
\arguments{
\item{...}{arguments passed to base function,
plus any of the \link{interactive_parameters}.}
}
\description{
This geometry is based on \code{\link[ggplot2:geom_dotplot]{ggplot2::geom_dotplot()}}.
See the documentation for those functions for more details.
}
\section{Details for interactive geom functions}{

The interactive parameters can be supplied with two ways:
\itemize{
\item As aesthetics with the mapping argument (via \code{\link[ggplot2:aes]{ggplot2::aes()}}).
In this way they can be mapped to data columns and apply to a set of geometries.

\item As plain arguments into the geom_*_interactive function.
In this way they can be set to a scalar value.
}
}

\examples{
# add interactive dot plots to a ggplot -------
library(ggplot2)
library(ggiraph)

p <- ggplot(mtcars, aes(x = mpg, fill = factor(cyl))) +
  geom_dotplot_interactive(
    aes(tooltip = row.names(mtcars)),
    stackgroups = TRUE,
    binwidth = 1,
    method = "histodot"
  )

x <- girafe(ggobj = p)
if (interactive()) {
  print(x)
}

gg_point = ggplot(
  data = mtcars,
  mapping = aes(
    x = factor(vs),
    fill = factor(cyl),
    y = mpg,
    tooltip = row.names(mtcars)
  )
) +
  geom_dotplot_interactive(
    binaxis = "y",
    stackdir = "center",
    position = "dodge"
  )

x <- girafe(ggobj = gg_point)
if (interactive()) {
  print(x)
}
}
\seealso{
\code{\link[=girafe]{girafe()}}
}
