% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_jitter_interactive.R
\name{geom_jitter_interactive}
\alias{geom_jitter_interactive}
\title{interactive jittered points}
\usage{
geom_jitter_interactive(...)
}
\arguments{
\item{...}{arguments passed to base geometry.}
}
\description{
The geometry is based on \code{\link[ggplot2]{geom_jitter}}.
See the documentation for those functions for more details.
}
\examples{
# add interactive paths to a ggplot -------
library(ggplot2)
library(ggiraph)

gg_jitter <- ggplot(mpg, aes(cyl, hwy,
                     tooltip = paste(manufacturer, model, year, trans, sep = "\\n")))+
  geom_jitter_interactive()

x <- girafe(ggobj = gg_jitter)
if( interactive() ) print(x)
}
\seealso{
\code{\link{girafe}}
}
