% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/annotation_raster_interactive.R,
%   R/facet_interactive.R, R/geom_abline_interactive.R,
%   R/geom_path_interactive.R, R/geom_polygon_interactive.R,
%   R/geom_rect_interactive.R, R/geom_bar_interactive.R,
%   R/geom_boxplot_interactive.R, R/geom_col_interactive.R,
%   R/geom_contour_interactive.R, R/geom_crossbar_interactive.R,
%   R/geom_curve_interactive.R, R/geom_density_2d_interactive.R,
%   R/geom_density_interactive.R, R/geom_dotplot_interactive.R,
%   R/geom_errorbar_interactive.R, R/geom_errorbarh_interactive.R,
%   R/geom_hex_interactive.R, R/geom_hline_interactive.R,
%   R/geom_label_interactive.R, R/geom_linerange_interactive.R,
%   R/geom_map_interactive.R, R/geom_point_interactive.R,
%   R/geom_pointrange_interactive.R, R/geom_quantile_interactive.R,
%   R/geom_raster_interactive.R, R/geom_ribbon_interactive.R,
%   R/geom_segment_interactive.R, R/geom_sf_interactive.R,
%   R/geom_smooth_interactive.R, R/geom_spoke_interactive.R,
%   R/geom_text_interactive.R, R/geom_tile_interactive.R,
%   R/geom_violin_interactive.R, R/geom_vline_interactive.R,
%   R/guide_bins_interactive.R, R/guide_colourbar_interactive.R,
%   R/guide_coloursteps_interactive.R, R/guide_legend_interactive.R
\name{GeomInteractive}
\alias{GeomInteractive}
\alias{GeomInteractiveRasterAnn}
\alias{FacetInteractiveWrap}
\alias{FacetInteractiveGrid}
\alias{GeomInteractiveAbline}
\alias{GeomInteractivePath}
\alias{GeomInteractiveLine}
\alias{GeomInteractiveStep}
\alias{GeomInteractivePolygon}
\alias{GeomInteractiveRect}
\alias{GeomInteractiveBar}
\alias{StatInteractiveBoxplot}
\alias{GeomInteractiveBoxplot}
\alias{GeomInteractiveCol}
\alias{GeomInteractiveContour}
\alias{GeomInteractiveContourFilled}
\alias{GeomInteractiveCrossbar}
\alias{GeomInteractiveCurve}
\alias{GeomInteractiveDensity2d}
\alias{GeomInteractiveDensity2dFilled}
\alias{GeomInteractiveDensity}
\alias{GeomInteractiveDotplot}
\alias{GeomInteractiveErrorbar}
\alias{GeomInteractiveErrorbarh}
\alias{GeomInteractiveHex}
\alias{GeomInteractiveHline}
\alias{GeomInteractiveLabel}
\alias{GeomInteractiveLinerange}
\alias{GeomInteractiveMap}
\alias{GeomInteractivePoint}
\alias{GeomInteractivePointrange}
\alias{GeomInteractiveQuantile}
\alias{GeomInteractiveRaster}
\alias{GeomInteractiveRibbon}
\alias{GeomInteractiveArea}
\alias{GeomInteractiveSegment}
\alias{GeomInteractiveSf}
\alias{GeomInteractiveSmooth}
\alias{GeomInteractiveSpoke}
\alias{GeomInteractiveText}
\alias{GeomInteractiveTile}
\alias{GeomInteractiveViolin}
\alias{GeomInteractiveVline}
\alias{GuideInteractiveBins}
\alias{GuideInteractiveColourbar}
\alias{GuideInteractiveColoursteps}
\alias{GuideInteractiveLegend}
\title{ggproto classes for ggiraph}
\description{
ggproto classes for ggiraph
}
\section{Geoms}{


All \verb{geom_*_interactive} functions (like \code{geom_point_interactive}) return a layer that
contains a \verb{GeomInteractive*} object (like \code{GeomInteractivePoint}). The \verb{Geom*}
object is responsible for rendering the data in the plot.

See \link[ggplot2:Geom]{Geom} for more information.
}

\keyword{internal}
