% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiraph.R
\name{ggiraph}
\alias{ggiraph}
\title{ggiraph}
\usage{
ggiraph(fun, graph.width = 7, graph.height = 7, pointsize = 12,
  width = NULL, height = NULL, ...)
}
\arguments{
\item{fun}{plot function. The function will be executed to produce graphics.
For \code{lattice} or \code{ggplot} object, the function should be \code{print}
and at least an extra argument \code{x} should specify the object
to plot. For traditionnal plots, the function should contain plot instructions. See examples.}

\item{graph.width}{plot width in inches (default value is 7).}

\item{graph.height}{plot height in inches (default value is 7).}

\item{pointsize}{the default pointsize of plotted text in pixels, default to 12.}

\item{width}{widget width}

\item{height}{widget height}

\item{...}{arguments for \code{fun}.}
}
\description{
Create an interactive graphic
to be used in a web browser.
}
\examples{
# ggiraph simple example -------
dataset = iris
dataset$tooltip = paste0( "Species <br/>", dataset$Species )
dataset$clickjs = paste0("function() {alert('",dataset$Species, "')}" )

# plots
gg_point_1 = ggplot(dataset, aes(x = Sepal.Length, y = Petal.Width, 
		color = Species, tooltip = tooltip, onclick=clickjs) ) + 
	geom_point_interactive(size=3)

ggiraph(fun=print, x = gg_point_1)
}

