% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplotwidget.R
\name{ggplotwidget}
\alias{ggplotwidget}
\title{ggplotwidget}
\usage{
ggplotwidget(fun, pointsize = 12, graph.width = 6, graph.height = 6,
  fontname = "Verdana", width = NULL, height = NULL, ...)
}
\arguments{
\item{fun}{fun}

\item{pointsize}{default to 12}

\item{graph.width}{gra w}

\item{graph.height}{graph h}

\item{fontname}{fn}

\item{width}{w}

\item{height}{h}

\item{...}{sddf sf}
}
\description{
<Add Description>
}

