% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplotwidget.R
\name{ggplotwidgetOutput}
\alias{ggplotwidgetOutput}
\title{Widget output function for use in Shiny}
\usage{
ggplotwidgetOutput(outputId)
}
\arguments{
\item{outputId}{outputId}
}
\description{
Widget output function for use in Shiny
}

