% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{girafe_defaults}
\alias{girafe_defaults}
\title{Get girafe defaults formatting properties}
\usage{
girafe_defaults(name = NULL)
}
\arguments{
\item{name}{optional, option's name to return, one of
'fonts', 'opts_sizing', 'opts_tooltip', 'opts_hover', 'opts_hover_key',
'opts_hover_inv', 'opts_hover_theme', 'opts_selection',
'opts_selection_inv', 'opts_selection_key', 'opts_selection_theme',
'opts_zoom', 'opts_toolbar'.}
}
\value{
a list containing default values or an
element selected with argument \code{name}.
}
\description{
The current formatting properties
are automatically applied to every girafe you produce.
These default values are returned by this function.
}
\examples{
girafe_defaults()
}
\seealso{
Other girafe animation options:
\code{\link[=girafe_options]{girafe_options()}},
\code{\link[=init_girafe_defaults]{init_girafe_defaults()}},
\code{\link[=opts_hover]{opts_hover()}},
\code{\link[=opts_selection]{opts_selection()}},
\code{\link[=opts_sizing]{opts_sizing()}},
\code{\link[=opts_toolbar]{opts_toolbar()}},
\code{\link[=opts_tooltip]{opts_tooltip()}},
\code{\link[=opts_zoom]{opts_zoom()}},
\code{\link[=set_girafe_defaults]{set_girafe_defaults()}}
}
\concept{girafe animation options}
