% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafe_export_svg.R
\name{girafe_export_svg}
\alias{girafe_export_svg}
\title{Export a girafe object to SVG}
\usage{
girafe_export_svg(x, filename, overwrite = FALSE)
}
\arguments{
\item{x}{A girafe object created via \code{\link[=girafe]{girafe()}}}

\item{filename}{The filename for the SVG document}

\item{overwrite}{Overwrite file if it already exists?}
}
\description{
Export a girafe object to a standalone SVG document.
}
\details{
The exported svg will include all necessary Javascript scripts and CSS styles,
so that it be fully interactive by itself, when viewed in a browser or
embedded in an HTML document via the \code{embed}, \code{object} or \code{iframe} tag.
}
\note{
\itemize{
\item The toolbar button 'save as png', is not supported inside the SVG.
\item If the svg document is used in an \code{img} tag, it will not be interactive.
}
}
\examples{
library(ggplot2)
library(ggiraph)
library(htmltools)

df <- mtcars
df$id <- rownames(mtcars)
p <- ggplot(df, aes(wt, mpg)) +
  geom_point_interactive(aes(tooltip = id, data_id = id), size = 5)
x <- girafe(
  ggobj = p,
  width_svg = 7,
  height_svg = 7,
  options = list(
    opts_zoom(min = 1, max = 4),
    opts_toolbar(saveaspng = TRUE),
    opts_hover(css = "fill:magenta;"),
    opts_hover_inv(css = "opacity:.3"),
    opts_selection(only_shiny = FALSE, css = "fill:red;"),
    opts_sizing(rescale = FALSE)
  )
)
workdir <- tempfile(pattern = "ggiraph")
dir.create(workdir)
svgfilename <- file.path(workdir, "test.svg")
girafe_export_svg(x, svgfilename, TRUE)
message("Exported SVG: ", svgfilename)

if (interactive()) {
  html <- tagList(
    tags$head(tags$title("ggiraph interactive standalone SVG")),
    tags$body(
      tags$p("This is an embedded standalone interactive svg"),
      tags$embed(
        type = "image/svg+xml",
        src = "test.svg"
      ),
      tags$div(
        tags$a(href = "test.svg", target = "_blank", "Click to open the svg in another tab")
      )
    )
  )
  htmlfilename <- file.path(workdir, "test.html")
  htmltools::save_html(
    html = html,
    file = htmlfilename
  )
  message("Exported HTML: ", htmlfilename)
  browseURL(htmlfilename)
} else {
  unlink(workdir, recursive = TRUE)
}
}
\seealso{
\code{\link[=girafe]{girafe()}}
}
