% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_colourbar_interactive.R
\name{guide_train.colourbar_interactive}
\alias{guide_train.colourbar_interactive}
\title{methods for interactive colourbar guide}
\usage{
\method{guide_train}{colourbar_interactive}(guide, scale,
  aesthetic = NULL)
}
\arguments{
\item{guide}{The guide object}

\item{scale, aesthetic}{other parameters used by guide_train}
}
\description{
These functions should not be used by the end users.
}
