% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_legend_interactive.R
\name{guide_train.legend_interactive}
\alias{guide_train.legend_interactive}
\title{methods for interactive legend guide}
\usage{
\method{guide_train}{legend_interactive}(guide, scale, aesthetic = NULL)
}
\arguments{
\item{guide}{The guide object}

\item{scale, aesthetic}{other parameters used by guide_train}
}
\description{
These functions should not be used by the end users.
}
