% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafe.R
\name{hover_settings}
\alias{hover_settings}
\title{hover effect}
\usage{
hover_settings(css = "fill:orange;stroke:gray;")
}
\arguments{
\item{css}{extra css used to customize animation.}
}
\description{
Allows customization of the animation
of graphic elements on which the mouse is positioned.
}
