% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeller_interactive.R
\name{id_label}
\alias{id_label}
\title{simple interactive label}
\usage{
id_label(x)
}
\arguments{
\item{x}{the input text}
}
\value{
an interactive label, see \code{\link[=label_interactive]{label_interactive()}}.
id_label("a")
}
\description{
A simple implementation to turn strip labels into
interactive strip labels.
}
\seealso{
\code{\link[=labeller_interactive]{labeller_interactive()}}
}
