% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{init_girafe_defaults}
\alias{init_girafe_defaults}
\title{Re-init animation defaults options}
\usage{
init_girafe_defaults()
}
\description{
Re-init all defaults options with the package defaults.
}
\seealso{
Other girafe animation options:
\code{\link[=girafe_defaults]{girafe_defaults()}},
\code{\link[=girafe_options]{girafe_options()}},
\code{\link[=opts_hover]{opts_hover()}},
\code{\link[=opts_selection]{opts_selection()}},
\code{\link[=opts_sizing]{opts_sizing()}},
\code{\link[=opts_toolbar]{opts_toolbar()}},
\code{\link[=opts_tooltip]{opts_tooltip()}},
\code{\link[=opts_zoom]{opts_zoom()}},
\code{\link[=set_girafe_defaults]{set_girafe_defaults()}}
}
\concept{girafe animation options}
