% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_path_grob.R
\name{interactive_path_grob}
\alias{interactive_path_grob}
\title{Generate interactive path grob}
\usage{
interactive_path_grob(..., tooltip = NULL, onclick = NULL,
  data_id = NULL, cl = "interactive_path_grob")
}
\arguments{
\item{...}{arguments passed to base function.}

\item{tooltip}{tooltip associated with paths}

\item{onclick}{javascript action to execute when path is clicked}

\item{data_id}{identifiers to associate with paths}

\item{cl}{class to set}
}
\description{
The grob is based on \code{\link[grid]{pathGrob}}.
See the documentation for that function for more details.
}
