% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_points_grob.R
\name{interactive_points_grob}
\alias{interactive_points_grob}
\title{Generate interactive points grob}
\usage{
interactive_points_grob(..., tooltip = NULL, onclick = NULL,
  data_id = NULL, cl = "interactive_points_grob")
}
\arguments{
\item{...}{arguments passed to base function.}

\item{tooltip}{tooltip associated with points}

\item{onclick}{javascript action to execute when point is clicked}

\item{data_id}{identifiers to associate with points}

\item{cl}{class to set}
}
\description{
The grob is based on \code{\link[grid]{pointsGrob}}.
See the documentation for that function for more details.
}
