% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_polygon_grob.R
\name{interactive_polygon_grob}
\alias{interactive_polygon_grob}
\title{Generate interactive polygon grob}
\usage{
interactive_polygon_grob(..., tooltip = NULL, onclick = NULL,
  data_id = NULL, cl = "interactive_polygon_grob")
}
\arguments{
\item{...}{arguments passed to base function.}

\item{tooltip}{tooltip associated with polygons}

\item{onclick}{javascript action to execute when polygon is clicked}

\item{data_id}{identifiers to associate with polygons}

\item{cl}{class to set}
}
\description{
The grob is based on \code{\link[grid]{polygonGrob}}.
See the documentation for that function for more details.
}
