% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_polyline_grob.R
\name{interactive_polyline_grob}
\alias{interactive_polyline_grob}
\title{Generate interactive polyline grob}
\usage{
interactive_polyline_grob(..., tooltip = NULL, onclick = NULL,
  data_id = NULL, cl = "interactive_polyline_grob")
}
\arguments{
\item{...}{arguments passed to base function.}

\item{tooltip}{tooltip associated with polylines}

\item{onclick}{javascript action to execute when polyline is clicked}

\item{data_id}{identifiers to associate with polylines}

\item{cl}{class to set}
}
\description{
The grob is based on \code{\link[grid]{polylineGrob}}.
See the documentation for that function for more details.
}
