% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_rect_grob.R
\name{interactive_rect_grob}
\alias{interactive_rect_grob}
\title{Generate interactive rectangle grob}
\usage{
interactive_rect_grob(..., tooltip = NULL, onclick = NULL,
  data_id = NULL, cl = "interactive_rect_grob")
}
\arguments{
\item{...}{arguments passed to base function.}

\item{tooltip}{tooltip associated with rectangles}

\item{onclick}{javascript action to execute when rectangle is clicked}

\item{data_id}{identifiers to associate with rectangles}

\item{cl}{class to set}
}
\description{
The grob is based on \code{\link[grid]{rectGrob}}.
See the documentation for that function for more details.
}
