% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_segments_grob.R
\name{interactive_segments_grob}
\alias{interactive_segments_grob}
\title{Generate interactive segments grob}
\usage{
interactive_segments_grob(..., tooltip = NULL, onclick = NULL,
  data_id = NULL, cl = "interactive_segments_grob")
}
\arguments{
\item{...}{arguments passed to base function.}

\item{tooltip}{tooltip associated with segments}

\item{onclick}{javascript action to execute when segment is clicked}

\item{data_id}{identifiers to associate with segments}

\item{cl}{class to set}
}
\description{
The grob is based on \code{\link[grid]{segmentsGrob}}.
See the documentation for that function for more details.
}
