% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_text_grob.R
\name{interactive_text_grob}
\alias{interactive_text_grob}
\title{Generate interactive text grob}
\usage{
interactive_text_grob(..., tooltip = NULL, onclick = NULL,
  data_id = NULL, cl = "interactive_text_grob")
}
\arguments{
\item{...}{arguments passed to base function.}

\item{tooltip}{tooltip associated with texts}

\item{onclick}{javascript action to execute when text is clicked}

\item{data_id}{identifiers to associate with texts}

\item{cl}{class to set}
}
\description{
The grob is based on \code{\link[grid]{textGrob}}.
See the documentation for that function for more details.
}
