% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interactivePolygonGrob.R
\name{interactivePolygonGrob}
\alias{interactivePolygonGrob}
\title{Generate interactive grob polygons}
\usage{
interactivePolygonGrob(x = unit(c(0, 1), "npc"), y = unit(c(0, 1), "npc"),
  id = NULL, id.lengths = NULL, tooltips = NULL, clicks = NULL,
  datid = NULL, default.units = "npc", name = NULL, gp = gpar(),
  vp = NULL)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-locations.}

\item{y}{A numeric vector or unit object specifying y-locations.}

\item{id}{A numeric vector used to separate locations in \code{x} and
    \code{y} into multiple polygons.  All locations with the same
    \code{id} belong to the same polygon.}

\item{id.lengths}{A numeric vector used to separate locations in \code{x} and
    \code{y} into multiple polygons.  Specifies consecutive blocks of
    locations which make up separate polygons.}

\item{tooltips}{tooltips associated with polygons}

\item{clicks}{javascript action to execute when polygon is clicked}

\item{datid}{identifiers to associate with polygons}

\item{default.units}{A string indicating the default units to use
    if \code{x}, \code{y}, \code{width}, or \code{height}
    are only given as numeric vectors.}

\item{name}{ A character identifier. }

\item{gp}{An object of class \code{gpar}, typically the output
    from a call to the function \code{gpar}.  This is basically
    a list of graphical parameter settings.}

\item{vp}{A Grid viewport object (or NULL).}
}
\description{
This function can be used to generate interactive grob
polygons.
}

