% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interactiveSegmentsGrob.R
\name{interactiveSegmentsGrob}
\alias{interactiveSegmentsGrob}
\title{Generate interactive grob segments}
\usage{
interactiveSegmentsGrob(x0 = unit(0, "npc"), y0 = unit(0, "npc"),
  x1 = unit(1, "npc"), y1 = unit(1, "npc"), tooltips = NULL,
  clicks = NULL, dbclicks = NULL, default.units = "npc", arrow = NULL,
  name = NULL, gp = gpar(), vp = NULL)
}
\arguments{
\item{tooltips}{tooltips associated with segments}

\item{clicks}{javascript action to execute when segment is clicked}

\item{dbclicks}{javascript action to execute when segment is double clicked}

\item{default.units}{A string indicating the default units to use
    if \code{x}, \code{y}, \code{width}, or \code{height}
    are only given as numeric vectors.}

\item{name}{ A character identifier. }

\item{gp}{An object of class \code{gpar}, typically the output
    from a call to the function \code{gpar}.  This is basically
    a list of graphical parameter settings.}

\item{vp}{A Grid viewport object (or NULL).}
}
\description{
This function can be used to generate interactive grob
segments.
}

