% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeller_interactive.R
\name{labeller_interactive}
\alias{labeller_interactive}
\title{Create a interactive strip label}
\usage{
labeller_interactive(
  custom_label_func,
  default_labeller = ggplot2::label_value
)
}
\arguments{
\item{custom_label_func}{a function that returns the result
of \code{\link[=label_interactive]{label_interactive()}}.}

\item{default_labeller}{a Labeller function, see \code{\link[=label_value]{label_value()}}}
}
\description{
These function is an helper function
to let user add interactivity on facet strip text.

The function adds interactive attributes but does not tell
to use them, make sure the theme you are using is specifying the use of \code{element_text_interactive}
for strip text,
for example \code{theme(strip.text = element_text_interactive())}
}
\examples{
library(ggplot2)
library(ggiraph)

gg_jitter <- ggplot(
  mpg, aes(cyl, hwy, group = cyl)) +
  geom_boxplot() +
  facet_wrap(~manufacturer,
             labeller = labeller_interactive(id_label)) +
  theme(strip.text = element_text_interactive())

x <- girafe(ggobj = gg_jitter)
if( interactive() ) print(x)
}
\seealso{
\code{\link[=id_label]{id_label()}}, \code{\link[=label_interactive]{label_interactive()}}
}
