% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_label_interactive.R
\name{makeContent.interactive_label_grob}
\alias{makeContent.interactive_label_grob}
\title{interactive_label_grob drawing}
\usage{
\method{makeContent}{interactive_label_grob}(x)
}
\arguments{
\item{x}{ A grid grob. }
}
\description{
draw an interactive_label_grob
}
