% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafe_options.R
\name{opt_tooltip}
\alias{opt_tooltip}
\title{settings for tooltip}
\usage{
opt_tooltip(css = NULL, offx = 10, offy = 0, opacity = 0.9,
  delay_mouseover = 200, delay_mouseout = 500, zindex = 999)
}
\arguments{
\item{css}{extra css (added to \code{position: absolute;pointer-events: none;})
used to customize tooltip area.}

\item{offx, offy}{tooltip x and y offset}

\item{opacity}{tooltip background opacity}

\item{delay_mouseover}{The duration in milliseconds of the
transition associated with tooltip display.}

\item{delay_mouseout}{The duration in milliseconds of the
transition associated with tooltip end of display.}

\item{zindex}{tooltip css z-index, default to 999.}
}
\description{
Settings to be used with \code{\link{girafe}}
for tooltip customisation.
}
\examples{
library(ggplot2)

dataset <- mtcars
dataset$carname = row.names(mtcars)

gg <- ggplot(
  data = dataset,
  mapping = aes(x = wt, y = qsec, color = disp,
                tooltip = carname, data_id = carname) ) +
  geom_point_interactive() + theme_minimal()

x <- girafe(ggobj = gg)
x <- girafe_options(x,
  opt_tooltip(opacity = .7,
    offx = 20, offy = -10,
    delay_mouseout = 1000) )
if( interactive() ) print(x)
}
\seealso{
set options with \code{\link{girafe_options}}

Other girafe animation options: \code{\link{opt_hover}},
  \code{\link{opt_selection}}, \code{\link{opt_zoom}}
}
\concept{girafe animation options}
