% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafe_options.R
\name{opts_hover}
\alias{opts_hover}
\alias{opts_hover_inv}
\alias{opts_hover_key}
\alias{opts_hover_theme}
\title{Hover effect settings}
\usage{
opts_hover(
  css = NULL,
  reactive = FALSE,
  nearest_distance = NULL,
  linked = FALSE
)

opts_hover_inv(css = NULL)

opts_hover_key(css = NULL, reactive = FALSE)

opts_hover_theme(css = NULL, reactive = FALSE)
}
\arguments{
\item{css}{css to associate with elements when they are hovered.
It must be a scalar character. It can also be constructed with
\code{\link[=girafe_css]{girafe_css()}}, to give more control over the css for different element types.}

\item{reactive}{if TRUE, in Shiny context, hovering will set Shiny input values.}

\item{nearest_distance}{a scalar positive number defining the maximum distance to use
when using the \code{hover_nearest} \link[=interactive_parameters]{interactive parameter} feature.
By default (\code{NULL}) it's set to \code{Infinity} which means that there is no distance limit.
Setting it to 50, for example, it will hover the nearest element that has
at maximum 50 SVG units (pixels) distance from the mouse cursor.}

\item{linked}{if TRUE, hovering a legend/guide element (\code{key-id}) will
also highlight the corresponding geometry elements (\code{data-id}) and vice versa.}
}
\description{
Allows customization of the rendering
of graphic elements when the user hovers over them with the cursor (mouse pointer).
Use \code{opts_hover} for interactive geometries in panels,
\code{opts_hover_key} for interactive scales/guides and
\code{opts_hover_theme} for interactive theme elements.
Use \code{opts_hover_inv} for the effect on the rest of the geometries,
while one is hovered (inverted operation).
}
\note{
\strong{IMPORTANT}: When applying a \code{fill} style with the \code{css} argument,
be aware that the browser's CSS engine will apply it also to line elements,
if there are any that use the hovering feature. This will cause an undesired effect.

To overcome this, supply the argument \code{css} using \code{\link[=girafe_css]{girafe_css()}},
in order to set the \code{fill} style only for the desired elements.
}
\examples{
library(ggplot2)

dataset <- mtcars
dataset$carname = row.names(mtcars)

gg <- ggplot(
  data = dataset,
  mapping = aes(x = wt, y = qsec, color = disp,
                tooltip = carname, data_id = carname) ) +
  geom_point_interactive() + theme_minimal()

x <- girafe(ggobj = gg)
x <- girafe_options(x,
  opts_hover(css = "fill:wheat;stroke:orange;r:5pt;") )
if( interactive() ) print(x)
}
\seealso{
\code{\link[=girafe_css]{girafe_css()}}, \code{\link[=girafe_css_bicolor]{girafe_css_bicolor()}}

Other girafe animation options:
\code{\link[=girafe_defaults]{girafe_defaults()}},
\code{\link[=girafe_options]{girafe_options()}},
\code{\link[=init_girafe_defaults]{init_girafe_defaults()}},
\code{\link[=opts_selection]{opts_selection()}},
\code{\link[=opts_sizing]{opts_sizing()}},
\code{\link[=opts_toolbar]{opts_toolbar()}},
\code{\link[=opts_tooltip]{opts_tooltip()}},
\code{\link[=opts_zoom]{opts_zoom()}},
\code{\link[=set_girafe_defaults]{set_girafe_defaults()}}
}
\concept{girafe animation options}
