% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafe_options.R
\name{opts_shiny_sizing}
\alias{opts_shiny_sizing}
\title{shiny sizing settings}
\usage{
opts_shiny_sizing(svg_auto_width = TRUE, svg_auto_height = FALSE,
  svg_limit_width = FALSE, svg_limit_height = FALSE)
}
\arguments{
\item{svg_auto_width}{When TRUE (default), sets the css width property to 100\%, otherwise it is set by the girafeOutput width parameter}

\item{svg_auto_height}{When TRUE, sets the css height property to 100\%, otherwise (default) it is set by the girafeOutput height parameter}

\item{svg_limit_width}{When TRUE, sets the css max-width property to the svg's width, otherwise (default) it is not set}

\item{svg_limit_height}{When TRUE, sets the css max-width property to the svg's height, otherwise (default) it is not set}
}
\description{
Allows customization of the svg style sizing in shiny
}
\seealso{
set options with \code{\link{girafe_options}}
}
\concept{girafe sizing options}
