% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafe_options.R
\name{opts_sizing}
\alias{opts_sizing}
\title{Girafe sizing settings}
\usage{
opts_sizing(rescale = TRUE, width = 1)
}
\arguments{
\item{rescale}{If FALSE, graphic will not be resized and the dimensions are exactly
those of the svg. If TRUE the graphic will be resize to fit its container}

\item{width}{widget width ratio (0 < width <= 1).}
}
\description{
Allows customization of the svg style sizing
}
\examples{
library(ggplot2)

dataset <- mtcars
dataset$carname = row.names(mtcars)

gg <- ggplot(
  data = dataset,
  mapping = aes(x = wt, y = qsec, color = disp,
                tooltip = carname, data_id = carname) ) +
  geom_point_interactive() + theme_minimal()

x <- girafe(ggobj = gg)
x <- girafe_options(x,
  opts_sizing(rescale = FALSE) )
if( interactive() ) print(x)
}
\seealso{
Other girafe animation options:
\code{\link[=girafe_defaults]{girafe_defaults()}},
\code{\link[=girafe_options]{girafe_options()}},
\code{\link[=init_girafe_defaults]{init_girafe_defaults()}},
\code{\link[=opts_hover]{opts_hover()}},
\code{\link[=opts_selection]{opts_selection()}},
\code{\link[=opts_toolbar]{opts_toolbar()}},
\code{\link[=opts_tooltip]{opts_tooltip()}},
\code{\link[=opts_zoom]{opts_zoom()}},
\code{\link[=set_girafe_defaults]{set_girafe_defaults()}}
}
\concept{girafe animation options}
