% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafe_options.R
\name{opts_zoom}
\alias{opts_zoom}
\title{Zoom settings}
\usage{
opts_zoom(min = 1, max = 1, duration = 300, default_on = FALSE)
}
\arguments{
\item{min}{minimum zoom factor}

\item{max}{maximum zoom factor}

\item{duration}{duration of the zoom transitions, in milliseconds}

\item{default_on}{if TRUE, pan/zoom will be activated by default when the plot is rendered}
}
\description{
Allows customization of the zoom.
}
\examples{
library(ggplot2)

dataset <- mtcars
dataset$carname = row.names(mtcars)

gg <- ggplot(
  data = dataset,
  mapping = aes(x = wt, y = qsec, color = disp,
                tooltip = carname, data_id = carname) ) +
  geom_point_interactive() + theme_minimal()

x <- girafe(ggobj = gg)
x <- girafe_options(x,
  opts_zoom(min = .7, max = 2, default_on = TRUE) )
if( interactive() ) print(x)
}
\seealso{
Other girafe animation options:
\code{\link[=girafe_defaults]{girafe_defaults()}},
\code{\link[=girafe_options]{girafe_options()}},
\code{\link[=init_girafe_defaults]{init_girafe_defaults()}},
\code{\link[=opts_hover]{opts_hover()}},
\code{\link[=opts_selection]{opts_selection()}},
\code{\link[=opts_sizing]{opts_sizing()}},
\code{\link[=opts_toolbar]{opts_toolbar()}},
\code{\link[=opts_tooltip]{opts_tooltip()}},
\code{\link[=set_girafe_defaults]{set_girafe_defaults()}}
}
\concept{girafe animation options}
