% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiraph.R
\name{renderggiraph}
\alias{renderggiraph}
\title{Reactive version of ggiraph object}
\usage{
renderggiraph(...)
}
\arguments{
\item{...}{unused}
}
\description{
Makes a reactive version of a ggiraph object for use in Shiny.

This function is now defunct, users should
now use function \code{\link[=renderGirafe]{renderGirafe()}}.
}
\keyword{internal}
