% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplotwidget.R
\name{renderggplotwidget}
\alias{renderggplotwidget}
\title{Widget render function for use in Shiny}
\usage{
renderggplotwidget(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{expr}

\item{env}{env}

\item{quoted}{quoted}
}
\description{
Widget render function for use in Shiny
}

