% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_shape_interactive.R
\name{scale_shape_interactive}
\alias{scale_shape_interactive}
\alias{scale_shape_continuous_interactive}
\alias{scale_shape_discrete_interactive}
\alias{scale_shape_binned_interactive}
\alias{scale_shape_ordinal_interactive}
\title{Create interactive scales for shapes}
\usage{
scale_shape_interactive(...)

scale_shape_continuous_interactive(...)

scale_shape_discrete_interactive(...)

scale_shape_binned_interactive(...)

scale_shape_ordinal_interactive(...)
}
\arguments{
\item{...}{arguments passed to base function,
plus any of the \link{interactive_parameters}.}
}
\value{
An interactive scale object.
}
\description{
These scales are based on
\code{\link[ggplot2:scale_shape]{ggplot2::scale_shape()}},
\code{\link[ggplot2:scale_shape_continuous]{ggplot2::scale_shape_continuous()}},
\code{\link[ggplot2:scale_shape_discrete]{ggplot2::scale_shape_discrete()}},
\code{\link[ggplot2:scale_shape_binned]{ggplot2::scale_shape_binned()}} and
\code{\link[ggplot2:scale_shape_ordinal]{ggplot2::scale_shape_ordinal()}}.
See the documentation for those functions for more details.
}
\section{Details for interactive scale and interactive guide functions}{

For scales, the interactive parameters can be supplied as arguments in the relevant function
and they can be scalar values or vectors, depending on the number of breaks (levels) and
the type of the guide used.
The guides do not accept any interactive parameter directly, they receive them from the scales.

When guide of type \code{legend}, \code{bins}, \code{colourbar} or \code{coloursteps} is used,
it will be converted to a \code{\link[=guide_legend_interactive]{guide_legend_interactive()}}, \code{\link[=guide_bins_interactive]{guide_bins_interactive()}},
\code{\link[=guide_colourbar_interactive]{guide_colourbar_interactive()}} or \code{\link[=guide_coloursteps_interactive]{guide_coloursteps_interactive()}} respectively,
if it's not already.

The length of each scale interactive parameter vector should match the length of the breaks.
It can also be a named vector, where each name should correspond to the same break name.
It can also be defined as function that takes the breaks as input and returns a named or
unnamed vector of values as output.

For binned guides like \code{bins} and \code{coloursteps} the breaks include the label breaks and the limits.
The number of bins will be one less than the number of breaks and the interactive parameters can be
constructed for each bin separately (look at the examples).
For \code{colourbar} guide in raster mode, the breaks vector, is scalar 1 always, meaning the interactive
parameters should be scalar too. For \code{colourbar} guide in non-raster mode, the bar is drawn using
rectangles, and the breaks are the midpoints of each rectangle.

The interactive parameters here, give interactivity only to the key elements of the guide.

To provide interactivity to the rest of the elements of a guide, (title, labels, background, etc),
the relevant theme elements or relevant guide arguments can be used.
The \code{guide} arguments \code{title.theme} and \code{label.theme} can be defined as
\code{element_text_interactive} (in fact, they will be converted to that if they are not
already), either directly or via the theme.
See the element_*_interactive section for more details.
}

\seealso{
\code{\link[=girafe]{girafe()}}

Other interactive scale:
\code{\link[=scale_alpha_interactive]{scale_alpha_interactive()}},
\code{\link[=scale_colour_brewer_interactive]{scale_colour_brewer_interactive()}},
\code{\link{scale_colour_interactive}},
\code{\link[=scale_colour_steps_interactive]{scale_colour_steps_interactive()}},
\code{\link{scale_gradient_interactive}},
\code{\link[=scale_linetype_interactive]{scale_linetype_interactive()}},
\code{\link{scale_manual_interactive}},
\code{\link[=scale_size_interactive]{scale_size_interactive()}},
\code{\link{scale_viridis_interactive}}
}
\concept{interactive scale}
