% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafe.R
\name{selection_settings}
\alias{selection_settings}
\title{hover effect}
\usage{
selection_settings(css = "fill:red;stroke:gray;", type = "multiple")
}
\arguments{
\item{css}{extra css used to customize animation.}

\item{type}{selection mode ("single", "multiple", "none")
when widget is in a Shiny application.}
}
\description{
Allows customization of the rendering of
selected graphic elements.
}
