% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafe.R
\name{zoom_settings}
\alias{zoom_settings}
\title{zoom settings}
\usage{
zoom_settings(min = 1, max = 1)
}
\arguments{
\item{min}{minimum zoom factor}

\item{max}{maximum zoom factor}
}
\description{
Allows customization of the zoom.
}
\examples{
library(ggplot2)

dataset <- mtcars
dataset$carname = row.names(mtcars)

gg <- ggplot(
  data = dataset,
  mapping = aes(x = wt, y = qsec, color = disp,
                tooltip = carname, data_id = carname) ) +
  geom_point_interactive() + theme_minimal()

x <- girafe(ggobj = gg)
x <- girafe_settings(x,
  zoom_settings(min = .7, max = 2) )
if( interactive() ) print(x)
}
