% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_contour_interactive.R
\name{geom_contour_interactive}
\alias{geom_contour_interactive}
\title{interactive 2d contours of a 3d surface.}
\usage{
geom_contour_interactive(...)
}
\arguments{
\item{...}{arguments passed to base geometry.}
}
\description{
The geometry is based on \code{\link[ggplot2]{geom_contour}}.
See the documentation for those functions for more details.
}
\examples{
# add interactive contours to a ggplot -------
library(ggplot2)
library(ggiraph)

v <- ggplot(faithfuld, aes(waiting, eruptions, z = density))
p <- v + geom_contour_interactive(aes(
  colour = stat(level),
  tooltip = paste("Level:", stat(level))
))
x <- girafe(ggobj = p)
if (interactive()) print(x)
}
\seealso{
\code{\link{girafe}}

\code{\link{girafe}}
}
