% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_histogram_interactive.R
\name{geom_histogram_interactive}
\alias{geom_histogram_interactive}
\title{interactive histogram}
\usage{
geom_histogram_interactive(...)
}
\arguments{
\item{...}{arguments passed to base geometry.}
}
\description{
The geometry is based on \code{\link[ggplot2]{geom_histogram}}.
See the documentation for those functions for more details.

This interactive version is only providing a single tooltip per
group of data (same for \code{data_id}). It means it is only possible
to associate a single tooltip to a set of bins.
}
\seealso{
\code{\link{girafe}}
}
