% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_polygon_interactive.R
\name{geom_polygon_interactive}
\alias{geom_polygon_interactive}
\title{add polygons with tooltips or click actions or double click actions}
\usage{
geom_polygon_interactive(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, see \code{\link[ggplot2]{geom_point}}.}

\item{data}{A data frame, see \code{\link[ggplot2]{geom_point}}.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string, see \code{\link[ggplot2]{geom_point}}.}

\item{position}{Postion adjustment, see \code{\link[ggplot2]{geom_point}}.}

\item{show.legend}{See \code{\link[ggplot2]{geom_point}}.}

\item{inherit.aes}{See \code{\link[ggplot2]{geom_point}}.}

\item{...}{other arguments passed on to layer. See \code{\link[ggplot2]{geom_point}}.}
}
\description{
Tooltips can be displayed when mouse is over polygons, on click actions and
double click actions can be set with javascript instructions.
}
\examples{
# add interactive polygons to a ggplot -------
ids <- factor(c("1.1", "2.1", "1.2", "2.2", "1.3", "2.3"))

values <- data.frame(
	id = ids,
	value = c(3, 3.1, 3.1, 3.2, 3.15, 3.5)
)

positions <- data.frame(
	id = rep(ids, each = 4),
	x = c(2, 1, 1.1, 2.2, 1, 0, 0.3, 1.1, 2.2, 1.1, 1.2, 2.5, 1.1, 0.3,
		0.5, 1.2, 2.5, 1.2, 1.3, 2.7, 1.2, 0.5, 0.6, 1.3),
	y = c(-0.5, 0, 1, 0.5, 0, 0.5, 1.5, 1, 0.5, 1, 2.1, 1.7, 1, 1.5,
		2.2, 2.1, 1.7, 2.1, 3.2, 2.8, 2.1, 2.2, 3.3, 3.2)
)

datapoly <- merge(values, positions, by=c("id"))

datapoly$value = as.character( datapoly$value )
datapoly$tooltips = as.character( datapoly$value )

gg_poly_1 <- ggplot(datapoly, aes( x = x, y = y ) ) + 
	geom_polygon_interactive(aes(fill = value, group = id, tooltips = tooltips))

ggplotwidget(fun=print, x = gg_poly_1)

}
\seealso{
\code{\link{geom_path_interactive}}, \code{\link{geom_point_interactive}}
}

