% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_raster_interactive.R
\name{geom_raster_interactive}
\alias{geom_raster_interactive}
\title{interactive raster rectangles.}
\usage{
geom_raster_interactive(...)
}
\arguments{
\item{...}{arguments passed to base geometry.}
}
\description{
The geometry is based on \code{\link[ggplot2]{geom_raster}}.
See the documentation for those functions for more details.
}
\examples{
# add interactive raster to a ggplot -------
library(ggplot2)
library(ggiraph)

df <- expand.grid(x = 0:5, y = 0:5)
df$z <- runif(nrow(df))

gg <- ggplot(df, aes(x, y, fill = z, tooltip = "tooltip")) +
  geom_raster_interactive() +
  scale_fill_gradient_interactive(
    data_id = "coco", onclick = "cici", tooltip = "cucu"
    )

x <- girafe(ggobj = gg)
if( interactive() ) print(x)
}
\seealso{
\code{\link{girafe}}

\code{\link{girafe}}
}
