% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_raster_interactive.R
\name{geom_raster_interactive}
\alias{geom_raster_interactive}
\title{interactive raster rectangles.}
\usage{
geom_raster_interactive(...)
}
\arguments{
\item{...}{arguments passed to base geometry.}
}
\description{
The geometry is based on \code{\link[ggplot2]{geom_raster}}.
See the documentation for those functions for more details.
}
\seealso{
\code{\link{girafe}}
}
