% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_text_interactive.R
\name{geom_text_interactive}
\alias{geom_text_interactive}
\title{add text with tooltips or click actions or double click actions}
\usage{
geom_text_interactive(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", parse = FALSE, ..., nudge_x = 0, nudge_y = 0,
  check_overlap = FALSE, na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{mapping}{The aesthetic mapping, see \code{\link[ggplot2]{geom_point}}.}

\item{data}{A data frame, see \code{\link[ggplot2]{geom_point}}.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string, see \code{\link[ggplot2]{geom_point}}.}

\item{position}{Postion adjustment, see \code{\link[ggplot2]{geom_point}}.}

\item{parse}{See \code{\link[ggplot2]{geom_point}}.}

\item{...}{other arguments passed on to layer. See \code{\link[ggplot2]{geom_point}}.}

\item{nudge_x, nudge_y}{See \code{\link[ggplot2]{geom_point}}.}

\item{check_overlap}{See \code{\link[ggplot2]{geom_point}}.}

\item{na.rm}{See \code{\link[ggplot2]{geom_point}}.}

\item{show.legend}{See \code{\link[ggplot2]{geom_point}}.}

\item{inherit.aes}{See \code{\link[ggplot2]{geom_point}}.}
}
\description{
tooltips can be displayed when mouse is over polygons, on click actions and
double click actions can be set with javascript instructions.
}
\examples{
# add interactive polygons to a ggplot -------
## the data
dataset = mtcars
dataset$label = row.names(mtcars)

dataset$tooltip = paste0( "<h5>", row.names(mtcars) ,"</h5>",
                          "<ul><li>cyl:", dataset$cyl, "</li>",
       "<li>gear:", dataset$gear, "</li>",
       "<li>carb:", dataset$carb, "</li></ul>")

## the plot
gg_text = ggplot(dataset,
                 aes(x = mpg, y = wt, label = label,
                     color = qsec,
                     tooltip = tooltip) ) +
  geom_text_interactive() +
  coord_cartesian(xlim = c(0,50))

## display the plot
ggiraph(fun=print, x = gg_text)
}
\seealso{
\code{\link{ggiraph}}
}

