% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/geom_path_interactive.R,
%   R/geom_polygon_interactive.R, R/geom_rect_interactive.R,
%   R/geom_bar_interactive.R, R/geom_boxplot_interactive.R,
%   R/geom_col_interactive.R, R/geom_hline_interactive.R,
%   R/geom_label_interactive.R, R/geom_map_interactive.R,
%   R/geom_point_interactive.R, R/geom_raster_interactive.R,
%   R/geom_segment_interactive.R, R/geom_sf_interactive.R,
%   R/geom_text_interactive.R, R/geom_tile_interactive.R,
%   R/geom_vline_interactive.R
\docType{data}
\name{GeomInteractive}
\alias{GeomInteractive}
\alias{GeomInteractivePath}
\alias{GeomInteractiveLine}
\alias{GeomInteractiveStep}
\alias{GeomInteractivePolygon}
\alias{GeomInteractiveRect}
\alias{GeomInteractiveBar}
\alias{GeomInteractiveBoxplot}
\alias{GeomInteractiveCol}
\alias{GeomInteractiveHline}
\alias{GeomInteractiveLabel}
\alias{GeomInteractiveMap}
\alias{GeomInteractivePoint}
\alias{GeomInteractiveRaster}
\alias{GeomInteractiveSegment}
\alias{GeomInteractiveSf}
\alias{GeomInteractiveText}
\alias{GeomInteractiveTile}
\alias{GeomInteractiveVline}
\title{ggproto classes for ggiraph}
\description{
ggproto classes for ggiraph
}
\section{Geoms}{


All `geom_*_interactive` functions (like `geom_point_interactive`) return a layer that
contains a `GeomInteractive*` object (like `GeomInteractivePoint`). The `Geom*`
object is responsible for rendering the data in the plot.

See \code{\link[ggplot2]{Geom}} for more information.
}

\keyword{datasets}
