% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiraph.R
\name{ggiraph}
\alias{ggiraph}
\title{Create a ggiraph object (defunct)}
\usage{
ggiraph(...)
}
\arguments{
\item{...}{unused}
}
\description{
Create an interactive graphic to be used in a web browser.

This function is now defunct, users should
now use function \code{\link[=girafe]{girafe()}}.
}
\keyword{internal}
