% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggiraph.R
\name{ggiraph}
\alias{ggiraph}
\title{ggiraph}
\usage{
ggiraph(fun, graph.width = 504, graph.height = 504, pointsize = 12,
  fontname = "Verdana", width = NULL, height = NULL, ...)
}
\arguments{
\item{fun}{plot function. The function will be executed to produce graphics.
For \code{lattice} or \code{ggplot} object, the function should be \code{print}
and at least an extra argument \code{x} should specify the object
to plot. For traditionnal plots, the function should contain plot instructions. See examples.}

\item{graph.width}{plot width in pixels (default value is 504).}

\item{graph.height}{plot height in pixels (default value is 504).}

\item{pointsize}{the default pointsize of plotted text in pixels, default to 12.}

\item{fontname}{the default font family to use, default to "Verdana".}

\item{width}{widget width}

\item{height}{widget height}

\item{...}{arguments for \code{fun}.}
}
\description{
Create an interactive graphic
to be used in a web browser.
}

