% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiraph.R
\name{ggiraphOutput}
\alias{ggiraphOutput}
\title{Create a ggiraph output element}
\usage{
ggiraphOutput(outputId)
}
\arguments{
\item{outputId}{output variable to read the ggiraph from.}
}
\description{
Render a ggiraph within an application page.
}
\examples{
if( require(shiny) && interactive() ){
  app_dir <- file.path( system.file(package = "ggiraph"), "shiny" )
  shinyAppDir(appDir = app_dir )
}
}

