% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiraph.R
\name{ggiraphOutput}
\alias{ggiraphOutput}
\title{Widget output function for use in Shiny}
\usage{
ggiraphOutput(outputId)
}
\arguments{
\item{outputId}{outputId}
}
\description{
Widget output function for use in Shiny
}

