% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafe_options.R
\name{girafe_options}
\alias{girafe_options}
\title{set girafe options}
\usage{
girafe_options(x, ...)
}
\arguments{
\item{x}{girafe object.}

\item{...}{set of options defined by calls to \code{opt_*} functions.}
}
\description{
Defines the animation options related to
a \code{\link{girafe}} object.
}
\examples{
library(ggplot2)

dataset <- mtcars
dataset$carname = row.names(mtcars)

gg_point = ggplot( data = dataset,
    mapping = aes(x = wt, y = qsec, color = disp,
    tooltip = carname, data_id = carname) ) +
  geom_point_interactive() + theme_minimal()

x <- girafe(ggobj = gg_point)
x <- girafe_options(x = x,
    opt_tooltip(opacity = .7),
    opt_zoom(min = .5, max = 4),
    opt_hover(css = "fill:red;stroke:orange;r:5pt;") )

if(interactive()){
  print(x)
}
}
\seealso{
\code{\link{opt_tooltip}}, \code{\link{opt_hover}},
\code{\link{opt_selection}}, \code{\link{opt_zoom}}
}
