% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafe.R
\name{girafeOutput}
\alias{girafeOutput}
\title{Create a girafe output element}
\usage{
girafeOutput(outputId, width = "100\%", height = "500px")
}
\arguments{
\item{outputId}{output variable to read the girafe from.}

\item{width}{widget width}

\item{height}{widget height}
}
\description{
Render a girafe within an application page.
}
