% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_colourbar_interactive.R
\name{guide_colourbar_interactive}
\alias{guide_colourbar_interactive}
\title{interactive colourbar guide}
\usage{
guide_colourbar_interactive(...)
}
\arguments{
\item{...}{arguments passed to guide_legend.}
}
\description{
an interactive legend guide.
See \code{\link[ggplot2]{guide_legend}} for more details.
}
