% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_legend_interactive.R
\name{guide_legend_interactive}
\alias{guide_legend_interactive}
\title{interactive legend guide}
\usage{
guide_legend_interactive(...)
}
\arguments{
\item{...}{arguments passed to guide_legend.}
}
\description{
an interactive legend guide.
See \code{\link[ggplot2]{guide_legend}} for more details.
}
