% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafe.R
\name{hover_settings}
\alias{hover_settings}
\title{hover effect}
\usage{
hover_settings(css = "fill:orange;stroke:gray;")
}
\arguments{
\item{css}{extra css used to customize animation.}
}
\description{
Allows customization of the animation
of graphic elements on which the mouse is positioned.
}
\examples{
library(ggplot2)

dataset <- mtcars
dataset$carname = row.names(mtcars)

gg <- ggplot(
  data = dataset,
  mapping = aes(x = wt, y = qsec, color = disp,
                tooltip = carname, data_id = carname) ) +
  geom_point_interactive() + theme_minimal()

x <- girafe(ggobj = gg)
x <- girafe_settings(x,
  hover_settings(css = "fill:wheat;stroke:orange;r:5pt;") )
if( interactive() ) print(x)
}
