% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveSegmentsGrob.R
\name{interactiveSegmentsGrob}
\alias{interactiveSegmentsGrob}
\title{Generate interactive grob segments}
\usage{
interactiveSegmentsGrob(x0 = unit(0, "npc"), y0 = unit(0, "npc"),
  x1 = unit(1, "npc"), y1 = unit(1, "npc"), tooltip = NULL,
  onclick = NULL, data_id = NULL, default.units = "npc", arrow = NULL,
  name = NULL, gp = gpar(), vp = NULL)
}
\arguments{
\item{x0}{ Numeric indicating the starting x-values of the line segments. }

\item{y0}{ Numeric indicating the starting y-values of the line
    segments. }

\item{x1}{ Numeric indicating the stopping x-values of the line
    segments. }

\item{y1}{ Numeric indicating the stopping y-values of the line segments.}

\item{tooltip}{tooltip associated with segments}

\item{onclick}{javascript action to execute when segment is clicked}

\item{data_id}{identifiers to associate with segments}

\item{default.units}{ A string. }

\item{arrow}{A list describing arrow heads to place at either end
    of the line segments, as produced by the \code{arrow} function.}

\item{name}{ A character identifier. }

\item{gp}{ An object of class \code{gpar}. }

\item{vp}{A Grid viewport object (or NULL).}
}
\description{
This function can be used to generate interactive grob
segments.
}

