% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_interactive.R
\name{label_interactive}
\alias{label_interactive}
\title{Create an interactive label}
\usage{
label_interactive(label, ...)
}
\arguments{
\item{label}{The text for the label (scalar character)}

\item{...}{any of the \code{\link[=interactive_parameters]{interactive_parameters()}}.}
}
\value{
an interactive label object
}
\description{
This function returns an object that can be used as a label
via the \code{\link[=labs]{labs()}} family of functions or
when setting a \code{scale}/\code{guide} name/title or key label.
It passes the interactive parameters to a theme element created via
\code{\link{element_text_interactive}} or via an interactive guide.
}
