% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{match_family}
\alias{match_family}
\title{Find best family match with systemfonts}
\usage{
match_family(font = "sans", bold = TRUE, italic = TRUE, debug = NULL)
}
\arguments{
\item{font}{family or face to match.}

\item{bold}{Wheter to match a font featuring a \code{bold} face.}

\item{italic}{Wheter to match a font featuring an \code{italic} face.}

\item{debug}{deprecated}
}
\description{
\code{match_family()} returns the best font family match.
}
\examples{
match_family("sans")
match_family("serif")
}
\seealso{
Other functions for font management: 
\code{\link{validated_fonts}()}
}
\concept{functions for font management}
