% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafe_options.R
\name{opts_sizing}
\alias{opts_sizing}
\title{girafe sizing settings}
\usage{
opts_sizing(rescale = TRUE, width = 1)
}
\arguments{
\item{rescale}{if TRUE, graphic will be resized when its container will be resized.}

\item{width}{widget width ratio (0 < width <= 1).}
}
\description{
Allows customization of the svg style sizing
}
\seealso{
set options with \code{\link{girafe_options}}
}
\concept{girafe sizing options}
