% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiraph.R
\name{renderggiraph}
\alias{renderggiraph}
\title{Widget render function for use in Shiny}
\usage{
renderggiraph(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{expr}

\item{env}{env}

\item{quoted}{quoted}
}
\description{
Widget render function for use in Shiny
}

