% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafe.R
\name{renderGirafe}
\alias{renderGirafe}
\title{Reactive version of girafe}
\usage{
renderGirafe(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns a \code{\link{}} object.}

\item{env}{The environment in which to evaluate expr.}

\item{quoted}{Is \code{expr} a quoted expression}
}
\description{
Makes a reactive version of a  object for use in Shiny.
}
