% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_viridis_interactive.R
\name{scale_viridis_interactive}
\alias{scale_viridis_interactive}
\alias{scale_colour_viridis_d_interactive}
\alias{scale_color_viridis_d_interactive}
\alias{scale_fill_viridis_d_interactive}
\alias{scale_colour_viridis_c_interactive}
\alias{scale_color_viridis_c_interactive}
\alias{scale_fill_viridis_c_interactive}
\title{Create interactive viridis colour scales}
\usage{
scale_colour_viridis_d_interactive(...)

scale_color_viridis_d_interactive(...)

scale_fill_viridis_d_interactive(...)

scale_colour_viridis_c_interactive(...)

scale_color_viridis_c_interactive(...)

scale_fill_viridis_c_interactive(...)
}
\arguments{
\item{...}{arguments passed to base function,
plus any of the \code{\link{interactive_parameters}}.}
}
\value{
An interactive scale object.
}
\description{
These scales are based on
\code{\link[ggplot2]{scale_colour_viridis_d}},
\code{\link[ggplot2]{scale_fill_viridis_d}},
\code{\link[ggplot2]{scale_colour_viridis_c}} and
\code{\link[ggplot2]{scale_fill_viridis_c}}.
See the documentation for those functions for more details.
}
\section{Details for scale_*_interactive functions}{

The interactive parameters can be supplied as arguments in the relevant function
and they can be scalar values or vectors, depending on the number of breaks (levels).
Their length should match the length of the breaks. They can also be
named vectors, where each name should correspond to the same break name.
Note that they give interactivity only to the key element of the legend.
For the title, the text labels and the background of the legend,
the relevant theme elements or relevant guide arguments
can be used to add interactivity. See the guide_\*_interactive and
element_\*_interactive sections below for more details.
}

\seealso{
\code{\link{girafe}}

Other interactive scale: \code{\link{scale_alpha_interactive}},
  \code{\link{scale_colour_brewer_interactive}},
  \code{\link{scale_colour_interactive}},
  \code{\link{scale_gradient_interactive}},
  \code{\link{scale_linetype_interactive}},
  \code{\link{scale_manual_interactive}},
  \code{\link{scale_shape_interactive}},
  \code{\link{scale_size_interactive}}
}
\concept{interactive scale}
