% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafe.R
\name{selection_settings}
\alias{selection_settings}
\title{hover effect}
\usage{
selection_settings(css = "fill:red;stroke:gray;", type = "multiple")
}
\arguments{
\item{css}{extra css used to customize animation.}

\item{type}{selection mode ("single", "multiple", "none")
when widget is in a Shiny application.}
}
\description{
Allows customization of the rendering of
selected graphic elements.
}
\examples{
library(ggplot2)

dataset <- mtcars
dataset$carname = row.names(mtcars)

gg <- ggplot(
  data = dataset,
  mapping = aes(x = wt, y = qsec, color = disp,
                tooltip = carname, data_id = carname) ) +
  geom_point_interactive() + theme_minimal()

x <- girafe(ggobj = gg)
x <- girafe_settings(x,
  selection_settings(type = "multiple",
    css = "fill:red;stroke:gray;r:5pt;") )
if( interactive() ) print(x)
}
