% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafe.R
\name{tooltip_settings}
\alias{tooltip_settings}
\title{settings for tooltip}
\usage{

  tooltip_settings(css = "padding:5px;background:black;color:white;border-radius:2px 2px 2px 2px;",
  offx = 10, offy = 0, opacity = 0.9, delay_mouseover = 200,
  delay_mouseout = 500, zindex = 999)
}
\arguments{
\item{css}{extra css (added to \code{position: absolute;pointer-events: none;})
used to customize tooltip area.}

\item{offx}{tooltip x offset}

\item{offy}{tooltip y offset}

\item{opacity}{tooltip opacity}

\item{delay_mouseover}{delai}

\item{delay_mouseout}{delai opacity}

\item{zindex}{tooltip css z-index, default to 999.}
}
\description{
settings to be used with \code{\link{girafe}}
for tooltip customisation.
}
