% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girafe.R
\name{zoom_settings}
\alias{zoom_settings}
\title{zoom settings}
\usage{
zoom_settings(min = 1, max = 1)
}
\arguments{
\item{min}{minimum zoom factor}

\item{max}{maximum zoom factor}
}
\description{
Allows customization of the zoom.
}
